[3/x] Add test page (#1716)
* [3/x] Add test page

* [4/x] pass key to menu content (#1717)

* [4/x] pass key to menu content

* [5/x] add page switcher, second page, and fallback page (#1718)

* [5/x] add page switcher, second page, and fallback page

* [6/x] placeholder components (#1719)

* [6/x] placeholder components

* Several forms decoupled (#1720)

* Several forms decoupled

* [8/x] add button to go to settings form (#1721)

* [8/x] add button to go to settings form

* [9/x] all handled by routing (#1722)

* [9/x] all handled by routing

* [11/x] move reconfigure (#1842)

* [11/x] move reconfigure

* [12/x] somewhere (#1843)
