+++ src/components/v2/V2Project/ArchiveV2Project.tsx
@@ -1,19 +1,15 @@
 import { V2ProjectContext } from 'contexts/v2/projectContext'
-import { TransactorInstance } from 'hooks/Transactor'
 import { useContext } from 'react'
 
 import ArchiveProject from 'components/ArchiveProject'
+import { useEditV2ProjectDetailsTx } from 'hooks/v2/transactor/EditV2ProjectDetailsTx'
 
 /**
  * V2 Wrapper around `ArchiveProject`.
  */
-export default function ArchiveV2Project({
-  editV2ProjectDetailsTx,
-}: {
-  editV2ProjectDetailsTx: TransactorInstance<{
-    cid: string
-  }>
-}) {
+export default function ArchiveV2Project() {
+  const editV2ProjectDetailsTx = useEditV2ProjectDetailsTx()
+
   const { projectMetadata, projectId, projectOwnerAddress, cv } =
     useContext(V2ProjectContext)
