+++ src/components/v2/V2Project/V2FundingCycleSection/ReservedTokensSplitsCard.tsx
@@ -17,6 +17,10 @@ import { tokenSymbolText } from 'utils/tokenSymbolText'
 import { formatReservedRate } from 'utils/v2/math'
 import { reloadWindow } from 'utils/windowUtils'
 
+import { pushSettingsContent } from 'utils/pushSettingsPage'
+
+import { useRouter } from 'next/router'
+
 import DistributeReservedTokensModal from './modals/DistributeReservedTokensModal'
 import { EditTokenAllocationModal } from './modals/EditTokenAllocationModal'
 
@@ -29,6 +33,7 @@ export default function ReservedTokensSplitsCard({
   reservedTokensSplits: Split[] | undefined
   reservedRate: BigNumber | undefined
 }) {
+  const router = useRouter()
   const { tokenSymbol, projectOwnerAddress, projectId, isPreviewMode } =
     useContext(V2ProjectContext)
   const {
@@ -164,7 +169,7 @@ export default function ReservedTokensSplitsCard({
             {canEditTokens && reservedRate?.gt(0) ? (
               <Button
                 size="small"
-                onClick={() => setEditTokenAllocationModalVisible(true)}
+                onClick={() => pushSettingsContent(router, 'reservedtokens')}
                 icon={<SettingOutlined />}
                 style={{ marginBottom: '1rem' }}
               >
