+++ src/components/v2/V2Project/V2ProjectHeaderActions.tsx
@@ -1,46 +1,26 @@
 import { t, Trans } from '@lingui/macro'
+import { SettingOutlined, ToolOutlined, HomeOutlined } from '@ant-design/icons'
 import { Button, Tooltip } from 'antd'
 import { V2ProjectToolsDrawer } from 'components/v2/V2Project/V2ProjectToolsDrawer/V2ProjectToolsDrawer'
 import { ThemeContext } from 'contexts/themeContext'
 import { V2ProjectContext } from 'contexts/v2/projectContext'
-import { useV2ConnectedWalletHasPermission } from 'hooks/v2/contractReader/V2ConnectedWalletHasPermission'
-import { V2OperatorPermission } from 'models/v2/permissions'
 
 import { useContext, useState } from 'react'
-import { ToolOutlined } from '@ant-design/icons'
 
 import ProjectVersionBadge from 'components/ProjectVersionBadge'
 
-import { featureFlagEnabled } from 'utils/featureFlags'
-
-import { InfoCircleOutlined } from '@ant-design/icons'
-
 import { useRouter } from 'next/router'
 
-import V2ReconfigureFundingModalTrigger from './V2ProjectReconfigureModal/V2ReconfigureModalTrigger'
-import { FEATURE_FLAGS } from 'constants/featureFlags'
-
 export default function V2ProjectHeaderActions() {
-  const { projectId } = useContext(V2ProjectContext)
   const router = useRouter()
+
+  const { projectId } = useContext(V2ProjectContext)
 
   const {
     theme: { colors },
   } = useContext(ThemeContext)
 
   const [toolDrawerVisible, setToolDrawerVisible] = useState<boolean>(false)
-
-  const canReconfigure = useV2ConnectedWalletHasPermission(
-    V2OperatorPermission.RECONFIGURE,
-  )
-
-  const showReconfigureButton = canReconfigure
-  const settingsPageEnabled = featureFlagEnabled(FEATURE_FLAGS.SETTINGS_PAGE)
-
-  const handleSettingsPageButtonClick = (e: React.MouseEvent<HTMLElement>) => {
-    e.preventDefault()
-    router.push(`/v2/p/${projectId}/settings`)
-  }
 
   return (
     <div
@@ -67,6 +47,28 @@ export default function V2ProjectHeaderActions() {
         onClose={() => setToolDrawerVisible(false)}
       />
       <div style={{ display: 'flex' }}>
+        <Tooltip title={t`Project Home`} placement="bottom">
+          <Button
+            onClick={() =>
+              router.push(`/v2/p/${projectId}`, undefined, {
+                shallow: true,
+              })
+            }
+            icon={<HomeOutlined />}
+            type="text"
+          />
+        </Tooltip>
+        <Tooltip title={t`Project Settings`} placement="bottom">
+          <Button
+            onClick={() =>
+              router.push(`/v2/p/${projectId}/settings`, undefined, {
+                shallow: true,
+              })
+            }
+            icon={<SettingOutlined />}
+            type="text"
+          />
+        </Tooltip>
         <Tooltip title={t`Tools`} placement="bottom">
           <Button
             onClick={() => setToolDrawerVisible(true)}
@@ -74,16 +76,6 @@ export default function V2ProjectHeaderActions() {
             type="text"
           />
         </Tooltip>
-        {showReconfigureButton && <V2ReconfigureFundingModalTrigger />}
-        {settingsPageEnabled && (
-          <Tooltip title={t`Open settings page`} placement="bottom">
-            <Button
-              icon={<InfoCircleOutlined />}
-              type="text"
-              onClick={handleSettingsPageButtonClick}
-            />
-          </Tooltip>
-        )}
       </div>
     </div>
   )
