+++ src/components/v2/V2Project/V2ProjectSettings/V2ProjectSettingsFundingCycleContent.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const V2ProjectSettingsFundingCycleContent = () => {
+  return <div>V2ProjectSettingsFundingCycleContent</div>
+}
+
+export default V2ProjectSettingsFundingCycleContent
