+++ src/components/v2/V2Project/V2ProjectSettings/V2ProjectSettingsPaymentAddressesContent.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const V2ProjectSettingsPaymentAddressesContent = () => {
+  return <div>V2ProjectSettingsPaymentAddressesContent</div>
+}
+
+export default V2ProjectSettingsPaymentAddressesContent
