+++ src/components/v2/V2Project/index.tsx
@@ -1,4 +1,4 @@
-import { Button, Space } from 'antd'
+import { Space } from 'antd'
 import ProjectHeader from 'components/Project/ProjectHeader'
 import { V2ProjectContext } from 'contexts/v2/projectContext'
 // TODO: Do we still need lazy loading?
@@ -11,19 +11,17 @@ import useProjectQueuedFundingCycle from 'hooks/v2/contractReader/ProjectQueuedF
 
 import { useIsUserAddress } from 'hooks/IsUserAddress'
 
-import { ProjectPage } from 'models/project-visibility'
+import { useRouter } from 'next/router'
+
+import { pushSettingsContent } from 'utils/pushSettingsPage'
 
 import { RelaunchFundingCycleBanner } from './banners/RelaunchFundingCycleBanner'
 import V2ProjectHeaderActions from './V2ProjectHeaderActions'
 
 import V2ProjectInfo from './V2ProjectInfo'
-import V2ProjectSettings from './V2ProjectSettingsPage/V2ProjectSettings'
 
-export default function V2Project({
-  selectedPage,
-}: {
-  selectedPage: ProjectPage
-}) {
+export default function V2Project() {
+  const router = useRouter()
   const {
     projectId,
     projectMetadata,
@@ -34,7 +32,6 @@ export default function V2Project({
     handle,
   } = useContext(V2ProjectContext)
 
-  const [activePage, setActivePage] = useState<ProjectPage>(selectedPage)
   const [handleModalVisible, setHandleModalVisible] = useState<boolean>()
 
   const canReconfigureFundingCycles = useV2ConnectedWalletHasPermission(
@@ -70,20 +67,15 @@ export default function V2Project({
         handle={handle}
         owner={projectOwnerAddress}
         onClickSetHandle={
-          showAddHandle ? () => setHandleModalVisible(true) : undefined
+          showAddHandle
+            ? () => pushSettingsContent(router, 'projecthandle')
+            : undefined
         }
       />
-      <Space direction="horizontal">
-        <Button onClick={() => setActivePage('info')}>Show Info</Button>
-        <Button onClick={() => setActivePage('settings')}>Show Settings</Button>
-      </Space>
-      {activePage === 'info' && (
-        <V2ProjectInfo
-          handleModalVisible={handleModalVisible}
-          setHandleModalVisible={setHandleModalVisible}
-        />
-      )}
-      {activePage === 'settings' && <V2ProjectSettings />}
+      <V2ProjectInfo
+        handleModalVisible={handleModalVisible}
+        setHandleModalVisible={setHandleModalVisible}
+      />
     </Space>
   )
 }
